requirements.txt
anyio==3.6.1
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.29.1
gurobipy==9.5.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
kiwisolver==1.3.2
lxml==4.8.0
matplotlib==3.5.1
multitasking==0.0.10
networkx==2.6.3
numpy==1.22.2
orjson==3.6.8
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
polygon==1.0.6
polygon-api-client==1.0.2
pygame==2.1.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
rfc3986==1.5.0
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
urllib3==1.26.9
websocket-client==1.3.2
websockets==10.3
yfinance==0.1.70
